+++ entregas/RecinosJose/Reto007/IngestaTipo.java
@@ -1,0 +1,24 @@
+package Reto007;
+
+class IngestaTipo {
+    private String tipo;
+    private Ingestas ingestas;
+
+    public IngestaTipo(String tipo, Ingestas ingestas) {
+        this.tipo = tipo;
+        this.ingestas = ingestas;
+    }
+
+    public String getTipo() {
+        return tipo;
+    }
+
+    public Ingestas getIngestas() {
+        return ingestas;
+    }
+
+    public String toString() {
+        return "Ingesta " + tipo + ": " + ingestas.toString();
+    }
+   
+}
